Safari-611.1.10/LayoutTests/http/tests/security/window-name-after-cross-origin-main-frame-navigation-expected.txt

CONSOLE MESSAGE: window.name should equal "" after a cross-origin main frame navigation.
CONSOLE MESSAGE: window.name equals "".
